database bug

+++ database/migrations/2019_10_12_000000_create_users_table.php
@@ -14,9 +14,9 @@ class CreateUsersTable extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->integer('idUsuario', true)->unsigned();
+            $table->bigIncrements('idUsuario', true)->unsigned();
             $table->string('CPF')->unique('CPF');
-            $table->integer('enderecoId')->unsined();
+            $table->unsignedBigInteger('enderecoId')->unsined();
             $table->foreign('enderecoId')->references('idEndereco')->on('addresses');
             $table->string('dataNascimento')->nullable();
             $table->string('email')->unique('email');
