+++ database/migrations/2019_10_17_003807_create_outboundsuppliers_table.php
@@ -14,8 +14,8 @@ class CreateOutboundsuppliersTable extends Migration
     public function up()
     {
         Schema::create('outboundsuppliers', function (Blueprint $table) {
-            $table->integer('entrada_saidas_idES')->unsined();
-            $table->integer('fornecedor_idFornecedor')-> unsined();
+            $table->unsignedBigInteger('entrada_saidas_idES')->unsined();
+            $table->unsignedBigInteger('fornecedor_idFornecedor')-> unsined();
             $table->foreign('entrada_saidas_idES')->references('idES')->on('inouts');
             $table->foreign('fornecedor_idFornecedor')->references('idFornecedor')->on('providers');
         });
